requirements.txt
PyQt5==5.15.4
pillow==8.3.1
pytesseract==0.3.8
pyautogui==0.9.53
requests==2.26.0